+++ application/classes/Core/Contract/HttpExceptionInterface.php
@@ -1,0 +1,10 @@
+<?php
+namespace Core\Contract {
+	interface HttpExceptionInterface {
+		/**
+		 * Return an integer representing a HTTP code
+		 * @return int Any HTTP code
+		 */
+		public function getHttpCode() : int;
+	}
+}