+++ application/classes/Core/Exception/Governance.php
@@ -1,0 +1,8 @@
+<?php
+namespace Core\Exception {
+	/**
+	 * \Core\Exception\Governance is raised when something should not be done
+	 * You should rarely have the need for catching this.
+	 */
+	class Governance extends \Exception {}
+}