+++ application/classes/Core/Exception/NotFound.php
@@ -1,13 +1,12 @@
 <?php
 namespace Core\Exception {
-	use \Exception;
-
-	class NotFound extends Exception {
+	/**
+	 * Throw this whenever you want to redirect the current controller to a not found controller
+	 */
+	class NotFound extends \Exception {
 		/**
-		 * Throw this whenever you want to redirect the current controller to a not found controller
+		 * @return int
 		 */
-		public function __construct(string $message = '', int $code = 0, ?Throwable $previous = null) {
-			parent::__construct($message, $code, $previous);
-		}
+		public function getHttpCode(): int { return 404; }
 	}
 }