+++ application/controllers/NotFound.php
@@ -1,18 +1,20 @@
 <?php
 	/**
-	 * Default controller for paths that cannot be routed.
-	 * @author Allan Thue Rehhoff
+	 * Controller for paths that either cannot be routed.
+	 * Or when a \Core\Exception\NotFound was thrown
 	 */
 	class NotFoundController extends Controller {
-
 		/**
 		 * Constructs an indicates the path wasn't found
-		 * @return void
 		 */
 		public function index() {
-			header("HTTP/1.0 404 Not Found");
+			if(IS_CLI) exit(1);
 
-			$this->setTitle("Not found");
-			$this->setView("notfound");
+			$this->response->data["httpHost"] = \HtmlEscape::escape($this->request->server["HTTP_HOST"] ?? '');
+			$this->response->data["requestUri"] = \HtmlEscape::escape($this->request->server["REQUEST_URI"] ?? '');
+			$this->response->data["httpReferer"] = \HtmlEscape::escape($this->request->server["HTTP_REFERER"] ?? '');
+
+			$this->response->setTitle("Not found");
+			$this->response->setView("notfound");
 		}
 	}